+++ app/cpt_file/parametrization.py
@@ -14,55 +14,179 @@ SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-from typing import List
 
-from viktor.parametrization import HiddenField
-from viktor.parametrization import LineBreak
-from viktor.parametrization import NumberField
-from viktor.parametrization import OptionField
-from viktor.parametrization import OptionListElement
-from viktor.parametrization import Parametrization
-from viktor.parametrization import Section
-from viktor.parametrization import SetParamsButton
-from viktor.parametrization import Tab
-from viktor.parametrization import Text
-from viktor.parametrization import TableInput
-from .constants import DEFAULT_MIN_LAYER_THICKNESS
+from viktor.parametrization import (
+    HiddenField,
+    Step,
+    DownloadButton,
+    BooleanField,
+    Lookup,
+    FileField,
+    IsEqual,
+    TextField,
+    LineBreak,
+    NumberField,
+    OptionField,
+    OptionListElement,
+    Parametrization,
+    SetParamsButton,
+    Text,
+    TableInput,
+    And,
+)
+from .constants import (
+    DEFAULT_MIN_LAYER_THICKNESS,
+    DEFAULT_CLASSIFICATION_TABLE,
+    DEFAULT_SOIL_NAMES,
+    MAX_CONE_RESISTANCE_TYPE,
+)
 from .constants import DEFAULT_ROBERTSON_TABLE
 
 
-def _get_soils_options(**kwargs) -> List[OptionListElement]:
-    """Options all possible soil type from the Classification parameter in the Project entity."""
-    return [OptionListElement(soil.get('ui_name')) for soil in DEFAULT_ROBERTSON_TABLE]
+CLASSIFICATION_METHODS = [
+    OptionListElement(label="Robertson Method (Fugro)", value="robertson"),
+    OptionListElement(label="Table Method", value="table"),
+]
 
 
 class CPTFileParametrization(Parametrization):
     """Defines the input fields in left-side of the web UI in the CPT_file entity (Editor)."""
-    gef = Tab('GEF')
-    gef.text = Text("Use the table below to change the interpreted soil layout by changing the positions of the layers, "
-                    "adding rows or changing the material type.")
 
-    gef.filter_thin_layers = SetParamsButton('Filter Layer Thickness',
-                                             method='filter_soil_layout_on_min_layer_thickness', flex=60,
-                                             description="Filter the soil layout to remove layers that are "
-                                                         "thinner than the minimum layer thickness")
-    gef.min_layer_thickness = NumberField('Minimum Layer Thickness', suffix='mm', min=0, step=50,
-                                          default=DEFAULT_MIN_LAYER_THICKNESS,
-                                          flex=40)
+    classification = Step("Upload and classification")
+    classification.text_01 = Text(
+        """# Welcome to the CPT interpretation app!
 
-    gef.reset_original_layers = SetParamsButton('Reset to original Soil Layout',
-                                                method='reset_soil_layout_user',
-                                                flex=100,
-                                                description="Reset the table to the original soil layout" )
+With this app you will be able to classify and interpret GEF-formatted CPT files by uploading and automatically 
+classifying the soil profile based on the classification table.
 
-    gef.ground_water_level = NumberField('Phreatic level', name='ground_water_level', suffix='m NAP', flex=50)
-    gef.ground_level = NumberField('Ground level', name='ground_level', suffix='m NAP', flex=50)
-    gef.soil_layout = TableInput('Soil layout', name='soil_layout')
-    gef.soil_layout.name = OptionField("Material", options=_get_soils_options)
-    gef.soil_layout.top_of_layer = NumberField("Top (m NAP)", num_decimals=1)
+## Step 1: Upload a GEF file
+
+For the users who want to try out the app, but do not have a GEF file at hand, feel free to use the sample 
+GEF file available.
+    """
+    )
+    classification.gef_file = FileField("Upload GEF file", file_types=[".gef"])
+    classification.get_sample_gef_toggle = BooleanField(
+        "Get sample GEF file", default=False, flex=15
+    )
+    classification.download_sample_gef = DownloadButton(
+        "Download sample GEF file",
+        "download_sample_gef_file",
+        visible=Lookup("classification.get_sample_gef_toggle"),
+        flex=15,
+    )
+    classification.text_02 = Text(
+        """## Step 2: Select your classification method
+        
+Select your preferred classification method.
+        """
+    )
+    classification.method = OptionField(
+        "Classification method",
+        options=CLASSIFICATION_METHODS,
+        default="robertson",
+        autoselect_single_option=True,
+        variant="radio-inline",
+        description="Robertson method: Robertson method, optimized for the dutch soil by Fugro. \n"
+        "\n Table method: Custom classification.",
+    )
+    classification.lb001 = LineBreak()
+    classification.alter_table = BooleanField("Alter classification table")
+    classification.robertson = TableInput(
+        "Robertson table",
+        default=DEFAULT_ROBERTSON_TABLE,
+        visible=And(
+            Lookup("classification.alter_table"),
+            IsEqual(Lookup("classification.method"), "robertson"),
+        ),
+    )
+    classification.robertson.name = TextField("Robertson Zone")
+    classification.robertson.ui_name = OptionField("Soil", options=DEFAULT_SOIL_NAMES)
+    classification.robertson.color = TextField("Color (R, G, B)")
+    classification.robertson.gamma_dry = NumberField("γ dry [kN/m³]", num_decimals=1)
+    classification.robertson.gamma_wet = NumberField("γ wet [kN/m³]", num_decimals=1)
+    classification.robertson.phi = NumberField("Friction angle Phi [°]", num_decimals=1)
+
+    classification.table = TableInput(
+        "Classification table",
+        default=DEFAULT_CLASSIFICATION_TABLE,
+        visible=And(
+            Lookup("classification.alter_table"),
+            IsEqual(Lookup("classification.method"), "table"),
+        ),
+    )
+    classification.table.name = OptionField("Naam", options=DEFAULT_SOIL_NAMES)
+    classification.table.color = TextField("Kleur (R, G, B)")
+    classification.table.qc_min = NumberField("qc min [MPa]", num_decimals=2)
+    classification.table.qc_max = NumberField("qc max [MPa]", num_decimals=2)
+    classification.table.qc_norm_min = NumberField("qc norm; min [MPa]", num_decimals=1)
+    classification.table.qc_norm_max = NumberField("qc norm; max [MPa]", num_decimals=1)
+    classification.table.rf_min = NumberField("Rf min [%]", num_decimals=1)
+    classification.table.rf_max = NumberField("Rf max [%]", num_decimals=1)
+    classification.table.gamma_dry = NumberField("γ dry [kN/m³]", num_decimals=1)
+    classification.table.gamma_wet = NumberField("γ wet [kN/m³]", num_decimals=1)
+    classification.table.phi = NumberField("Friction angle Phi [°]", num_decimals=1)
+    classification.table.max_cone_res_type = OptionField(
+        "Maximum cone resistance type", options=MAX_CONE_RESISTANCE_TYPE
+    )
+    classification.table.max_cone_res_mpa = NumberField(
+        "Maximum cone resistance [MPa]", num_decimals=1
+    )
+    classification.text_03 = Text(
+        """## Step 3: Classify the soil layout
+        
+Classify the uploaded GEF file by clicking the "Classify soil layout" button. Proceed then to the next step.
+        """
+    )
+    classification.classify_soil_layout_button = SetParamsButton(
+        "Classify soil layout", "classify_soil_layout"
+    )
+
+    cpt = Step("CPT interpretation", views=["visualize_cpt", "visualize_map"])
+    cpt.text = Text(
+        "Use the table below to change the interpreted soil layout by changing the positions of the layers, "
+        "adding rows or changing the material type."
+    )
+
+    cpt.filter_thin_layers = SetParamsButton(
+        "Filter Layer Thickness",
+        method="filter_soil_layout_on_min_layer_thickness",
+        flex=60,
+        description="Filter the soil layout to remove layers that are "
+        "thinner than the minimum layer thickness",
+    )
+    cpt.min_layer_thickness = NumberField(
+        "Minimum Layer Thickness",
+        suffix="mm",
+        min=0,
+        step=50,
+        default=DEFAULT_MIN_LAYER_THICKNESS,
+        flex=40,
+    )
+
+    cpt.reset_original_layers = SetParamsButton(
+        "Reset to original Soil Layout",
+        method="reset_soil_layout_user",
+        flex=100,
+        description="Reset the table to the original soil layout",
+    )
+
+    cpt.ground_water_level = NumberField(
+        "Phreatic level", name="ground_water_level", suffix="m NAP", flex=50
+    )
+    cpt.ground_level = NumberField(
+        "Ground level", name="ground_level", suffix="m NAP", flex=50
+    )
+    cpt.soil_layout = TableInput("Soil layout", name="soil_layout")
+    cpt.soil_layout.name = OptionField("Material", options=DEFAULT_SOIL_NAMES)
+    cpt.soil_layout.top_of_layer = NumberField("Top (m NAP)", num_decimals=1)
 
     # hidden fields
-    gef.gef_headers = HiddenField('GEF Headers', name='headers')
-    gef.bottom_of_soil_layout_user = HiddenField('GEF Soil bottom', name='bottom_of_soil_layout_user')
-    gef.measurement_data = HiddenField('GEF Measurement data', name='measurement_data')
-    gef.soil_layout_original = HiddenField('Soil layout original', name='soil_layout_original')
+    cpt.gef_headers = HiddenField("GEF Headers", name="headers")
+    cpt.bottom_of_soil_layout_user = HiddenField(
+        "GEF Soil bottom", name="bottom_of_soil_layout_user"
+    )
+    cpt.measurement_data = HiddenField("GEF Measurement data", name="measurement_data")
+    cpt.soil_layout_original = HiddenField(
+        "Soil layout original", name="soil_layout_original"
+    )