+++ CHANGELOG.md
@@ -3,10 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
-None.
+- Table method interpretation
+- Step editor
 
 ### Changed
 - translated all ui to english
+- Filefield upload instead of File-like entity type for GEF file.
 
 ### Deprecated
 None.
@@ -16,7 +18,7 @@ None.
 
 ### Fixed
 - update to simple app type
-- update to sdk 13.5
+- update to sdk 13.6.2
 
 ## v1.0.0 (2022/02/21)
 ### Added